+++ api/server/services/Config/getCustomConfig.js
@@ -1,4 +1,4 @@
-const { CacheKeys } = require('librechat-data-provider');
+const { CacheKeys, EModelEndpoint } = require('librechat-data-provider');
 const loadCustomConfig = require('./loadCustomConfig');
 const getLogStores = require('~/cache/getLogStores');
 
@@ -22,4 +22,19 @@ async function getCustomConfig() {
   return customConfig;
 }
 
-module.exports = getCustomConfig;
+/**
+ *
+ * @param {string | EModelEndpoint} endpoint
+ */
+const getCustomEndpointConfig = async (endpoint) => {
+  const customConfig = await getCustomConfig();
+  if (!customConfig) {
+    throw new Error(`Config not found for the ${endpoint} custom endpoint.`);
+  }
+
+  const { endpoints = {} } = customConfig;
+  const customEndpoints = endpoints[EModelEndpoint.custom] ?? [];
+  return customEndpoints.find((endpointConfig) => endpointConfig.name === endpoint);
+};
+
+module.exports = { getCustomConfig, getCustomEndpointConfig };
